[DM-17872] Parsing for math and coercing types

+++ cadc-adql/src/test/java/ca/nrc/cadc/tap/parser/TestUtil.java
@@ -131,9 +131,9 @@ public class TestUtil
         td = new TableDesc(sn, tn);
         sd.getTableDescs().add(td);
         td.getColumnDescs().add( createColumnDesc(tn, "t_integer", TapDataType.INTEGER, "int column", null, null, null));
-        td.getColumnDescs().add( new ColumnDesc(tn, "t_long", TapDataType.INTEGER));
-        td.getColumnDescs().add( new ColumnDesc(tn, "t_float", TapDataType.INTEGER));
-        td.getColumnDescs().add( new ColumnDesc(tn, "t_double", TapDataType.INTEGER));
+        td.getColumnDescs().add( new ColumnDesc(tn, "t_long", TapDataType.LONG));
+        td.getColumnDescs().add( new ColumnDesc(tn, "t_float", TapDataType.FLOAT));
+        td.getColumnDescs().add( new ColumnDesc(tn, "t_double", TapDataType.DOUBLE));
         td.getColumnDescs().add( new ColumnDesc(tn, "t_char", new TapDataType("char", "8", null)));
         td.getColumnDescs().add( createColumnDesc(tn, "t_varchar", new TapDataType("char", "8*", null), "varchar column", null, null, null));
         td.getColumnDescs().add( new ColumnDesc(tn, "t_string", new TapDataType("char", "8*", null)));
